+++ src/controllers/levels/xp/xp.ts
@@ -1,52 +1,49 @@
-import * as Re from '../../services/response';
-import {getConnection} from '../../services/connection';
+import * as Re from '../../../services/response';
+import {getConnection} from '../../../services/connection';
 
 const con = getConnection();
-module Chests {
+module XP {
     /**
-     * @api {post} /levels/chests/:userid Add Chests
+     * @api {post} /levels/xp/:userid  Add XP
      * @apiVersion 1.2.2
-     * @apiName AddChests
-     * @apiDescription Add Chests
+     * @apiName AddXP
+     * @apiDescription Add xp (windows xp?)
      * @apiGroup Levels
      *
      * @apiParam {String} username
      * @apiParam {Object} avatar
-     * @apiParam {Number} chests
-     * @apiParam {Object} discriminator
+     * @apiParam {Number} xp
+     * @apiParam {Object} discriminator Seems like a number, but who cares
      *
      * @apiSuccess (200) {Object} result
      * @apiSuccess (200) {String} result.action
-     * @apiSuccess (200) {Number} result.oldChests
-     * @apiSuccess (200) {Number} result.newChests
+     * @apiSuccess (200) {Number} result.oldXP
+     * @apiSuccess (200) {Number} result.newXP
      *
      * @apiSuccessExample {json} Success-Example:
      *     HTTP/1.1 200
      *     {
      *        "action": "add",
-     *        "oldChests": 3
-     *        "newChests": 5
+     *        "oldXP": 3
+     *        "newXP": 5
      *     }
      *
-     * @apiError userid_too_long Userid can only be 18 characters long (500 code for some reason)
-     * @apiError body_missing Request Body is missing (500 code for some reason)
-     * @apiError property_required Property name required (400 for some reason)
      *
      */
 
-    export async function addChest(req, res) {
+   export async function addXP (req, res) {
         const userid = req.params.userid;
         if (userid.length > 18) { return Re.userid_too_long(res); }
         const body = req.body;
         if (!body) { return Re.body_missing(res); }
 
-        const needed = ['username', 'avatar', 'chests', 'discriminator'];
+        const needed = ['username', 'avatar', 'xp', 'discriminator'];
         for (let i = 0; i < needed.length; i++) {
             if (!body.hasOwnProperty(needed[i])) {
                 return Re.property_required(res, needed[i]);
             }
         }
-        con.query('UPDATE discord_levels SET chests = chests + ? WHERE userid = ?', [body.chests, userid], function (error, results) {
+        con.query('UPDATE discord_levels SET xp = xp + ? WHERE userid = ?', [body.xp, userid], function (error, results) {
             if (error) { throw error; }
             if (results.changedRows === 0) {
                 let data = {
@@ -54,36 +51,36 @@ module Chests {
                     username: body.username,
                     discriminator: body.discriminator,
                     avatar: body.avatar,
-                    xp: 0,
-                    chests: body.chests,
+                    xp: body.xp,
+                    chests: 0,
                     coins: 100
                 };
                 con.query('INSERT INTO discord_levels SET ?', [data], function (error) {
                     if (error) { throw error; }
                 });
             }
-            con.query('SELECT chests FROM discord_levels WHERE userid = ?', [userid], function (error, results) {
+            con.query('SELECT xp FROM discord_levels WHERE userid = ?', [userid], function (error, results) {
                 if (error) { throw error; }
-                const newChests = results[0].chests;
+                const newXP = results[0].xp;
                 Re.success(res, {
                     action: 'add',
-                    oldChests: newChests - body.chests,
-                    newChests: newChests
+                    oldXP: newXP - body.xp,
+                    newXP: newXP
                 });
             });
         });
     }
 
     /**
-     * @api {delete} /levels/chests/:userid Delete Chests
+     * @api {delete} /levels/xp/:userid Delete XP
      * @apiVersion 1.2.2
-     * @apiName DeleteChests
-     * @apiDescription Delete Chests
+     * @apiName DeleteXP
+     * @apiDescription Delete XP
      * @apiGroup Levels
      *
      * @apiParam {String} username
      * @apiParam {Object} avatar
-     * @apiParam {Number} chests
+     * @apiParam {Number} xp
      * @apiParam {Object} discriminator
      *
      * @apiSuccess (200) {Object} result
@@ -95,28 +92,24 @@ module Chests {
      *        "action": "delete"
      *     }
      *
-     * @apiError userid_too_long Userid can only be 18 characters long (500 code for some reason)
-     * @apiError not_sufficientv User has not enough chests (500 for some reason)
-     * @apiError body_missing Request Body is missing (500 code for some reason)
-     * @apiError property_required Property name required (400 for some reason)
      *
      */
-    export async function deleteChest(req, res) {
+   export async function deleteXP (req, res) {
         const userid = req.params.userid;
         if (userid.length > 18) { return Re.userid_too_long(res); }
-
         const body = req.body;
         if (!body) { return Re.body_missing(res); }
 
-        const needed = ['username', 'avatar', 'chests', 'discriminator'];
+        const needed = ['username', 'avatar', 'xp', 'discriminator'];
         for (let i = 0; i < needed.length; i++) {
-            if (!body.hasOwnProperty(needed[i])) { return Re.property_required(res, needed[i]); }
+            if (!body.hasOwnProperty(needed[i])) {
+                return Re.property_required(res, needed[i]);
+            }
         }
-
-        con.query('SELECT chests FROM discord_levels WHERE userid = ?', [userid], function (error, results) {
+        con.query('SELECT xp FROM discord_levels WHERE userid = ?', [userid], function (error, results) {
             if (error) { throw error; }
-            const chests = results[0];
-            if (!chests) {
+            const xp = results[0];
+            if (!xp) {
                 // USER isn't in DB yet
                 let data = {
                     userid: userid,
@@ -129,20 +122,20 @@ module Chests {
                 };
                 con.query('INSERT INTO discord_levels SET ?', [data], function (error) {
                     if (error) { throw error; }
-                    return Re.not_sufficient(res, 'chests');
+                    return Re.not_sufficient(res, 'xp');
                 });
-            } else if (chests.chests >= body.chests) {
-                con.query('UPDATE discord_levels SET chests = chests - ? WHERE userid = ?', [body.chests, userid], function (error) {
+            } else if (xp.cp >= body.xp) {
+                con.query('UPDATE discord_levels SET xp = xp - ? WHERE userid = ?', [body.xp, userid], function (error) {
                     if (error) { throw error; }
                     Re.success(res, {
                         action: 'delete'
                     });
                 });
             } else {
-                // USER has not enough CHESTS
-                Re.not_sufficient(res, 'chests');
+                // USER has not enough XP
+                return Re.not_sufficient(res, 'xp');
             }
         });
     }
 }
-export = Chests;
+export = XP;
