+++ src/routes/donate.ts
@@ -1,11 +1,12 @@
-import {Request, Response, Router} from 'express';
+import {Router} from 'express';
 import {donate} from '../services/route';
+const DonationSchema = require('../models/validation/donation/donation');
 
 const Donation = require('../controllers/donation/donation');
 
 let donate_router = Router();
 
-donate_router.post(donate('/'), Donation.addDonation); // Add Donation
+donate_router.post(donate('/'), Vayder.validateBody(DonationSchema), Donation.addDonation); // Add Donation
 donate_router.get(donate('/'), Donation.findAllDonations);
 
 export default donate_router;
